swagger docs fur Konto

+++ noti/src/main/java/com/duvi/services/noti/service/impl/NotiServiceImpl.java
@@ -1,13 +1,14 @@
 package com.duvi.services.noti.service.impl;
 
-import com.duvi.services.noti.client.AccountClient;
+import  com.duvi.services.noti.client.AccountClient;
 import com.duvi.services.noti.client.StatsClient;
 import com.duvi.services.noti.domain.NotificationEntity;
 import com.duvi.services.noti.domain.Recipient;
 import com.duvi.services.noti.domain.dto.AccountDTO;
 import com.duvi.services.noti.domain.dto.NotificationDTO;
-import com.duvi.services.noti.domain.dto.StatsDTO;
 import com.duvi.services.noti.domain.enums.NotiType;
+import com.duvi.services.noti.domain.exception.NotificationAlreadyExists;
+import com.duvi.services.noti.domain.exception.NotificationNotExists;
 import com.duvi.services.noti.repository.NotiRepository;
 import com.duvi.services.noti.service.EmailService;
 import com.duvi.services.noti.service.NotiService;
@@ -16,10 +17,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
-import reactor.core.publisher.Flux;
 
+import java.io.IOException;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
 
 @Service
@@ -39,24 +39,33 @@ public class NotiServiceImpl implements NotiService {
     }
 
     @Override
-    public NotificationEntity createNotification(NotificationDTO dto) {
+    public NotificationEntity createNotification(NotificationDTO dto) throws NotificationAlreadyExists {
         NotificationEntity createdNoti = new NotificationEntity(dto);
+        if (notiRepository.existsByAccountNameAndNotiType(dto.accountName(), dto.notiType())) {
+            throw new NotificationAlreadyExists(dto.accountName(), dto.notiType());
+        }
         return notiRepository.save(createdNoti);
     }
 
     @Override
-    public List<NotificationEntity> getNotificationsByName(String accountName) {
+    public List<NotificationEntity> getNotificationsByName(String accountName) throws NotificationNotExists {
+        if (!notiRepository.existsByAccountName(accountName)) {
+            throw new NotificationNotExists(accountName);
+        }
         return notiRepository.findAllByAccountName(accountName);
     }
 
     @Override
-    public NotificationEntity deleteNotification(NotificationDTO dto) {
-        return notiRepository.deleteByAccountNameAndNotiType(dto.accountName(), dto.notiType());
+    public void deleteNotification(NotificationDTO dto) throws NotificationNotExists {
+        if (!notiRepository.existsByAccountNameAndNotiType(dto.accountName(), dto.notiType())) {
+           throw new NotificationNotExists(dto.accountName(), dto.notiType());
+        }
+        notiRepository.deleteByAccountNameAndNotiType(dto.accountName(), dto.notiType());
     }
 
     @Override
     @Scheduled(cron = "${remind.cron}")
-    public void sendRemindNotifications() throws MessagingException {
+    public Integer sendRemindNotifications() throws MessagingException {
 
         List<NotificationEntity> notifications = notiRepository.findAllRemindReady();
 
@@ -70,16 +79,18 @@ public class NotiServiceImpl implements NotiService {
                         noti.setLastNotified(LocalDate.now());
                         notiRepository.save(noti);
                         logger.info("Remind notification sent to {} successfully", noti.getEmail());
-                    } catch (MessagingException e) {
+                    } catch (MessagingException | IOException e) {
                         logger.info("Exception while sending remind notification to {}", noti.getEmail());
                     }
                 }
         );
+
+        return notifications.size();
     }
 
     @Override
     @Scheduled(cron = "${backup.cron}")
-    public void sendBackupNotifications() throws MessagingException {
+    public Integer sendBackupNotifications() throws MessagingException {
         List<NotificationEntity> notifications = notiRepository.findAllBackupReady();
 
         logger.info("found {} backup notifications ready to be sent!", notifications.size());
@@ -93,10 +104,14 @@ public class NotiServiceImpl implements NotiService {
                         noti.setLastNotified(LocalDate.now());
                         notiRepository.save(noti);
                         logger.info("Backup notification sent to {} successfully", noti.getEmail());
-                    } catch (MessagingException e) {
+                    } catch (MessagingException | IOException e) {
                         logger.info("Exception while sending backup notification to {}", noti.getEmail());
                     }
                 }
         );
+
+        return notifications.size();
+
+
     }
 }
